requirements.txt
Django==2.2.24
djangorestframework==3.11.2
Pillow==8.2.0
psycopg2==2.8.6
pytz==2019.3
sqlparse==0.2.4
